LayoutTests/fast/xpath/nsresolver-exception-expected.txt

CONSOLE MESSAGE: line 14: this is an exception
SUCCESS
